anticipating on null cases

+++ figis-search-web/src/main/java/org/figis/search/web/resource/IndexResource.java
@@ -14,6 +14,9 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
+import org.apache.commons.lang3.StringUtils;
+import org.fao.fi.factsheetwebservice.domain.FactsheetDomain;
+import org.figis.search.config.elements.Index;
 import org.figis.search.config.ref.FigisSearchException;
 import org.figis.search.service.IndexResponse;
 import org.figis.search.service.IndexResponse.OperationStatus;
@@ -45,10 +48,15 @@ public class IndexResource {
 	@Produces(MediaType.APPLICATION_JSON)
 	@ApiOperation(value = "Update the index of a specific domain", response = IndexResponse.class)
 	public IndexResponse updateDomain(
-			@PathParam("index") @ApiParam(value = "the name of the index", required = true) String index,
+			@PathParam("index") @ApiParam(value = "the name of the index", required = true) Index index,
 			@PathParam("action") @ApiParam(value = "action", required = true) Action action,
-			@PathParam("domain") @ApiParam(value = "the factsheet domain", required = true) String domain) {
-		return processAction(action, () -> service.update(index, domain), () -> service.delete(index, domain));
+			@PathParam("domain") @ApiParam(value = "the factsheet domain", required = true) FactsheetDomain domain) {
+
+		if (index == null || action == null || domain == null) {
+			return provideFailed();
+		} else {
+			return processAction(action, () -> service.update(index, domain), () -> service.delete(index, domain));
+		}
 	}
 
 	@GET
@@ -56,15 +64,23 @@ public class IndexResource {
 	@Produces(MediaType.APPLICATION_JSON)
 	@ApiOperation(value = "Update the index of a specific factsheet from a certain domain", response = IndexResponse.class)
 	public IndexResponse updateFactsheet(
-			@PathParam("index") @ApiParam(value = "the name of the index", required = true) String index,
+			@PathParam("index") @ApiParam(value = "the name of the index", required = true) Index index,
 			@PathParam("action") @ApiParam(value = "action", required = true) Action action,
-			@PathParam("domain") @ApiParam(value = "the factsheet domain", required = true) String domain,
+			@PathParam("domain") @ApiParam(value = "the factsheet domain", required = true) FactsheetDomain domain,
 			@PathParam("factsheet") @ApiParam(value = "the factsheet", required = true) String factsheet) {
-		return processAction(action, () -> service.update(index, domain, factsheet),
-				() -> service.delete(index, domain, factsheet));
+
+		if (index == null || action == null || domain == null || StringUtils.isBlank(factsheet)) {
+			return provideFailed();
+		} else {
+			return processAction(action, () -> service.update(index, domain, factsheet),
+					() -> service.delete(index, domain, factsheet));
+		}
 	}
 
 	private IndexResponse processAction(Action action, Callable<IndexResponse> update, Callable<IndexResponse> delete) {
+		if (action == null) {
+			action = Action.unknown;
+		}
 		try {
 			IndexResponse r;
 			switch (action) {
@@ -74,17 +90,23 @@ public class IndexResource {
 			case delete:
 				r = delete.call();
 				break;
+			case unknown:
 			default:
-				r = new IndexResponse();
-				r.setOperationStatus(OperationStatus.FAILED);
-				r.setMessageList(new ArrayList<String>());
-				r.getMessageList().add("No valid action specified");
+				r = provideFailed();
 				break;
 			}
 			return r;
 		} catch (Exception e) {
 			throw new FigisSearchException(e);
 		}
+	}
+
+	private IndexResponse provideFailed() {
+		IndexResponse r = new IndexResponse();
+		r.setOperationStatus(OperationStatus.FAILED);
+		r.setMessageList(new ArrayList<String>());
+		r.getMessageList().add("No valid action specified");
+		return r;
 	}
 
 }
